JAMES-1902 Extract 'data' cassandra implementation into its own maven module
