+++ server/container/guice/data-cassandra/src/main/java/org/apache/james/modules/data/CassandraRecipientRewriteTableModule.java
@@ -16,38 +16,47 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
+
 package org.apache.james.modules.data;
 
 import org.apache.james.backends.cassandra.components.CassandraModule;
+import org.apache.james.rrt.api.AliasReverseResolver;
+import org.apache.james.rrt.api.CanSendFrom;
+import org.apache.james.rrt.api.RecipientRewriteTable;
+import org.apache.james.rrt.cassandra.CassandraMappingsSourcesDAO;
+import org.apache.james.rrt.cassandra.CassandraRRTModule;
+import org.apache.james.rrt.cassandra.CassandraRecipientRewriteTable;
+import org.apache.james.rrt.cassandra.CassandraRecipientRewriteTableDAO;
+import org.apache.james.rrt.lib.AliasReverseResolverImpl;
+import org.apache.james.rrt.lib.CanSendFromImpl;
 import org.apache.james.server.core.configuration.ConfigurationProvider;
-import org.apache.james.user.api.UsersRepository;
-import org.apache.james.user.cassandra.CassandraUsersDAO;
-import org.apache.james.user.lib.UsersDAO;
-import org.apache.james.user.lib.UsersRepositoryImpl;
 import org.apache.james.utils.InitializationOperation;
 import org.apache.james.utils.InitilizationOperationBuilder;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Scopes;
-import com.google.inject.TypeLiteral;
 import com.google.inject.multibindings.Multibinder;
 import com.google.inject.multibindings.ProvidesIntoSet;
 
-public class CassandraUsersRepositoryModule extends AbstractModule {
+public class CassandraRecipientRewriteTableModule extends AbstractModule {
     @Override
     public void configure() {
-        bind(CassandraUsersDAO.class).in(Scopes.SINGLETON);
-        bind(UsersDAO.class).to(CassandraUsersDAO.class);
-        bind(new TypeLiteral<UsersRepositoryImpl<CassandraUsersDAO>>() {}).in(Scopes.SINGLETON);
-        bind(UsersRepository.class).to(new TypeLiteral<UsersRepositoryImpl<CassandraUsersDAO>>() {});
+        bind(CassandraRecipientRewriteTable.class).in(Scopes.SINGLETON);
+        bind(CassandraRecipientRewriteTableDAO.class).in(Scopes.SINGLETON);
+        bind(CassandraMappingsSourcesDAO.class).in(Scopes.SINGLETON);
+        bind(RecipientRewriteTable.class).to(CassandraRecipientRewriteTable.class);
+        bind(AliasReverseResolverImpl.class).in(Scopes.SINGLETON);
+        bind(AliasReverseResolver.class).to(AliasReverseResolverImpl.class);
+        bind(CanSendFromImpl.class).in(Scopes.SINGLETON);
+        bind(CanSendFrom.class).to(CanSendFromImpl.class);
         Multibinder<CassandraModule> cassandraDataDefinitions = Multibinder.newSetBinder(binder(), CassandraModule.class);
-        cassandraDataDefinitions.addBinding().toInstance(org.apache.james.user.cassandra.CassandraUsersRepositoryModule.MODULE);
+        cassandraDataDefinitions.addBinding().toInstance(CassandraRRTModule.MODULE);
     }
 
     @ProvidesIntoSet
-    InitializationOperation configureUsersRepository(ConfigurationProvider configurationProvider, UsersRepositoryImpl<CassandraUsersDAO> usersRepository) {
+    InitializationOperation configureRecipientRewriteTable(ConfigurationProvider configurationProvider, CassandraRecipientRewriteTable recipientRewriteTable) {
         return InitilizationOperationBuilder
-            .forClass(UsersRepositoryImpl.class)
-            .init(() -> usersRepository.configure(configurationProvider.getConfiguration("usersrepository")));
+            .forClass(CassandraRecipientRewriteTable.class)
+            .init(() -> recipientRewriteTable.configure(configurationProvider.getConfiguration("recipientrewritetable")));
     }
 }
